move StringOperations as it is used in different packages now
